Initial firebase integration
This is the initial firebase integration. It includes the framework for
version tracking, but that's not yet implemented on the front end. It
replaces ?code= with ?id= and no longer uses bitly. Of course, bitly can
be reintegrated if desired. It has also replaced the hard coded
sharepoint base URL with the window.location.href. This is for testing
purposes and can be changed back later. It currently does not work with
the tab functionality, ie. version tracking across tabs, etc.